Add and instrcution

+++ operations/jmp.go
@@ -19,7 +19,7 @@ func (jmp *JMP) Analyze(ctx *Context, inst byte) (int, string) {
 		retAddrStr := fmt.Sprintf("%04x", uint16(ctx.Idx+3)+retAddr16b)
 		return 3, getResult(ctx.Idx, getOrgOpe(ctx.Body[ctx.Idx:ctx.Idx+3]), getOpeString("jmp", retAddrStr))
 	case 0xeb:
-		retAddrStr := fmt.Sprintf("short %04x", byte(ctx.Idx+2)+ctx.Body[ctx.Idx+1])
+		retAddrStr := fmt.Sprintf("short %04x", (ctx.Idx+2)+int(ctx.Body[ctx.Idx+1]))
 		return 2, getResult(ctx.Idx, getOrgOpe(ctx.Body[ctx.Idx:ctx.Idx+2]), getOpeString("jmp", retAddrStr))
 	default:
 		return 100, ""
